Rename Rust options module to snuba_options
Address review feedback: the module previously named `options` was renamed to
`snuba_options` to avoid confusion with the `sentry_options` crate it imports
from (`use sentry_options::options`). Updates the module declaration, its one
importer, and a stale doc-comment reference.
